backend/pyproject.toml
[tool.poetry]
name = "example_backend"
version = "0.1.0"
description = "An example application using Flask, a database, Docker, etc"
authors = ["Stephen Mullins <smullins7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "7.1.2"
Flask = "1.1.2"
itsdangerous= "1.1.0"
Jinja2 = "2.11.3"
MarkupSafe = "1.1.1"
Werkzeug = "1.0.1"
Flask-SQLAlchemy = "2.5.1"
Flask-Migrate = "2.7.0"
SQLAlchemy-serializer = "^1.3.4"
Flask-Cors = "^3.0.10"
PyMySQL = "^1.0.2"
importlib-metadata = {version = "^3.4.0", python = "<3.8"}

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pytest-cov = "^2.11.1"
black = "20.8b1"  # match to .pre-commit-config.yaml
isort = "5.7.0"  # match to .pre-commit-config.yaml
pre-commit = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "debug"
addopts = "-v"
testpaths = ["tests"]

[tool.black]
line_length = 120

[tool.isort]
known_firstparty = ["example_backend"]
known_tests = ["tests"]
sections = ["FUTURE", "STDLIB","THIRDPARTY","TESTS","FIRSTPARTY","LOCALFOLDER"]
line_length = 120
profile = "black"

[tool.coverage.run]
source = ["example_backend"]

[tool.coverage.report]
fail_under = 75
